kano-toolset: Adding the kanoapplib.py
The kanoapplib is a basic python library for applications and projects
that contains some of the shared code. Currently, that is the WebApp
base class, that contains the python browser API that can be used to
develop web apps without having a local server server.

A part of kano-window-tool was moved to the kano app lib, because it
is useful in wider context.