kano-dialog: Making buttons respond to 'release'
Dialog buttons used to respond to button-press instead of release.
However, it is more common that actions are tied to release, so you can
cancel the action by moving away with your mouse pressed.

Another problem is that if we combine button press with button release,
clicking on a button in a dialog can trigger an action on a widget under
it, which is really weird. I spend about an hour debugging this today in
kano-apps ...