Improve error handling and remove assert()s.
Some caveats:

It's not obvious what to do if you start a dispmanx update and one of the intermediate api calls
failed. I'm assuming you have to finish an update if you started one.

The imageLayer convention for returning errors is true=success, which is the opposite to dispmanx
where 0=success. I have kept this.

libpng has a strange api involving longjmp and one call to deallocate everything.