Network: Make network_info report unaltered ESSID
KanoComputing/peldins#1832
Currently, to report the ESSID, `iwconfig` is queried and the output is
passed through grep, awk and sed. During this process, everything after
a space is scrapped and some of the other characters are lost. Instead
use Python's regex module to handle this to retain the spaces and
special characters.