Logging: Fix flush and sync functions
When `flush()`ing the log, it would close the file if one existed,
rather than flushing the file. It would also flush the `stderr`.

`sync`, on the other hand, would `flush()` the log file regardless of
whether it was open but would not flush the `stderr`.

Merge these two functions but retain both interfaces, ensuring that the
log file is open before flushing, flush the `stderr` and don't close the
file in any circumstance.