libkano: Fix package build issues
The `libkano` library requires the `libkano-python` library which is
provided as a subpackage of this repository. This means that the
libraries can't be an installed build dependency and so must be required
in such a fashion. To circumvent this problem, make the libraries'
development packages header install paths match that which is local to
the repository so that the repository path (and thus the include path)
can be identical to that when it is installed.