Prevent GUI thread from being blocked by UDHCPC hook processes
When connecting to WiFi, the kano.script is executed by the UDHCPC client.
In this script, we introduce various system hooks to the "internet connection
up" event. The latest one was kano-dashboard-sysupdates which launches the
Updater using the run_cmd command. Even though sysupdates is launched in the
background, processes spawned via run_cmd will block due to inheriting fds.

This fix says that the sysupdates should be run in the background and also
discard the standard output and error, on which it would block due to the
processes launched through run_cmd (since it uses Popen with piping
stdout and stderr).