Fix for comparison between str and unicode
Turns out that gtk will return a str when calling get_label. This breaks the
comparison with the _() function which returns a unicode when it successfully
was installed. Specifically, this fix is for all Kano OS dialogs which depend
on the output returned by the dialogs run() function. Review with care.