Parson: Fix i18n case where decimal separator is ','
trello.com/c/blt0FDXx/392-updater-status-file-is-incorrectly-parsed-in-spanish-due-to-the-decimal-separator
If LC_NUMERIC is set to a locale which uses `,` as the decimal separator
for numbers then the `strtod()` function will interpret `,`s following
integers as part of the number. This causes problems when integers are
in a JSON as there will be a `,` following the number to correctly
adhere to JSON syntax but this will be interpreted as part of the number
which causes there to be no `,` separator between keys.

For example, in

{
    "key_1": 10,
    "key_2": 20
}

`key_1` would parse `10,` as 10.0 and absorb the trailing `,`. Allow for
this case by checking for trailing `,`.