WebApp: Add Kano URI scheme comprehension
To interact with Kano World, a custom URI scheme was created to allow
launching of apps related to shares and the like. This scheme was not
implemented in the WebApp browser so any routing to one of these schemes
would lead to a 404 error. Prevent this happening by teaching it about
this scheme.