File Utils: Make sed() command more resistant
If the `sed()` command exported by the utils library is cut short for
some reason (e.g. if the power is pulled) then it will have overwritten
the file that it was supposed to be `sed`ing but not populated it.
Prevent this from happening by using a temporary file and copying it to
the destination.