Gtk: Migrate styling for Gtk 3.20
Changes in Gtk as of version 3.20 meant that elements are now styled via
CSS selectors which are slightly different from their class names; for
example, a `GtkButton` is now styled in CSS as `button` and a
`GtkMenuItem` is `menuitem`. Switch to use this new convention for all
our styling to allow windows to look as intended once again.