Logging: Fix logging tracebacks
When a program traces back, the logger will attempt to log it. However,
it is stored as a JSON in an intermediate form. This causes the
`json.dumps()` method to be invoked which, unfortunately, doesn't know
how to encode `Exception`s and thus traces back itself causing the
original traceback to be lost. Resolve this problem by casting
beforehand.