Fixed an issue with is_internet failing due to network iface name change
In Stretch, the ifaces now use predictable naming conventions and
have moved away from wlan0 or eth0. If the kernel option net.ifnames=0
is missing from a Stretch image, these hardcoded iface name tests might
fail.

It was observed that on an upgraded 3.14.0 image that the script
failed when using ethernet and the kernel option was not present since
eth iface was renamed but the wlan one was not. As such, connecting
to WiFi made the script succeed.